tests: move batch statements tests to batch.rs module
New batch statement test cases are moved to exising batch.rs test suite.

Additionally, the test cases for LWT and counter tables are unskipped
and adjusted to use keyspace with tablets disabled. This allows executing
the underlying test scenarios even though LWT and counters are not yet
supported with tablets.
