+++ scylla/tests/integration/main.rs
@@ -1,5 +1,4 @@
 mod batch;
-mod batch_statements;
 mod consistency;
 mod cql_collections;
 mod cql_types;
